backend/scheduler/requirements.txt
openai
python-dotenv
google-cloud-firestore